+++ README.md
@@ -116,13 +116,13 @@ The following flags are available for `registryCredentials delete`:
 * Delete a specific registry credential for a specific account `ocm support delete registryCredentials [accountID] [registryCredentialID]`
 * Delete all registry credentials for a specific account `ocm support delete registryCredentials [accountID] --all`
 
-### Find
+### Get
 
-The `find` command finds the given resource.
+The `get` command gets the given resource.
 
-#### Finding an account
+#### Getting an account
 
-Use the `accounts` subcommand to find one or more accounts, passing as argument one of the following:
+Use the `accounts` subcommand to get one or more accounts, passing as argument one of the following:
 
 * accountID
 * username
@@ -131,7 +131,7 @@ Use the `accounts` subcommand to find one or more accounts, passing as argument 
 * organizationExternalID
 * organizationEBSAccountID
 
-The following flags are available for `accounts find`:
+The following flags are available for `accounts get`:
 
 ```
 --all                        If true, returns all accounts that matched the search instead of the first one only (default behaviour).
@@ -139,18 +139,18 @@ The following flags are available for `accounts find`:
 --fetchRoles                 If true, includes the account roles.
 --fetchLabels                If true, includes the account labels.
 --fetchCapabilities          If true, includes the account capabilities.
--h, --help                   help for find
+-h, --help                   help for get
 ```
 
-#### Finding an organization
+#### Getting an organization
 
-Use the `organizations` subcommand to find one or more organizations, passing as argument one of the following:
+Use the `organizations` subcommand to get one or more organizations, passing as argument one of the following:
 
 * organizationID
 * organizationExternalID
 * organizationEBSAccountID
 
-The following flags are available for `organizations find`:
+The following flags are available for `organizations get`:
 
 ```
 --all                  If true, returns all organizations that matched the search instead of the first one only (default behaviour).
@@ -158,23 +158,23 @@ The following flags are available for `organizations find`:
 --fetchSubscriptions   If true, includes the organization subscriptions.
 --fetchLabels          If true, includes the organization labels.
 --fetchCapabilities    If true, includes the organization capabilities.
--h, --help             help for find
+-h, --help             help for get
 ```
 
-#### Finding registry credentials
+#### Getting registry credentials
 
-Use the `registryCredentials` subcommand to to find registry credentials, passing accountID.
+Use the `registryCredentials` subcommand to to get registry credentials, passing accountID.
 
 #### Examples
 
-* Find an account by email `ocm support find accounts user@redhat.com`
-* Find an account and include its roles in the results `ocm support find accounts [accountID] --fetchRoles`
-* Find an account and include its labels in the results `ocm support find accounts [accountID] --fetchLabels`
-* Find an account and include its capabilities in the results `ocm support find accounts [accountID] --fetchCapabilities`
-* Find all accounts from an organization `ocm support find accounts [organizationID] --all`
-* Find an organization by its externalID: `ocm support find organizations [organizationExternalID]`
-* Find an organization and include its subscriptions: `ocm support find organizations [organizationID] --fetchSubscriptions`
-* Find an organization and include its labels `ocm support find organizations [organizationID] --fetchLabels --fetchCapabilities`
-* Find an organization and include its capabilities `ocm support find organizations [organizationID] --fetchCapabilities`
-* Find an organization and include its quota: `ocm support find organizations [organizationID] --fetchQuota`
-* Show registry credentials for a specific account `ocm support find registryCredentials [accountID]`
+* Get an account by email `ocm support get accounts user@redhat.com`
+* Get an account and include its roles in the results `ocm support get accounts [accountID] --fetchRoles`
+* Get an account and include its labels in the results `ocm support get accounts [accountID] --fetchLabels`
+* Get an account and include its capabilities in the results `ocm support get accounts [accountID] --fetchCapabilities`
+* Get all accounts from an organization `ocm support get accounts [organizationID] --all`
+* Get an organization by its externalID: `ocm support get organizations [organizationExternalID]`
+* Get an organization and include its subscriptions: `ocm support get organizations [organizationID] --fetchSubscriptions`
+* Get an organization and include its labels `ocm support get organizations [organizationID] --fetchLabels --fetchCapabilities`
+* Get an organization and include its capabilities `ocm support get organizations [organizationID] --fetchCapabilities`
+* Get an organization and include its quota: `ocm support get organizations [organizationID] --fetchQuota`
+* Show registry credentials for a specific account `ocm support get registryCredentials [accountID]`
