+++ cmd/ocm-support/create/accountcapability/cmd.go
@@ -1,4 +1,4 @@
-package subscription_capability
+package accountcapability
 
 import (
 	"fmt"
@@ -8,32 +8,32 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/utils"
+	"github.com/openshift-online/ocm-support-cli/pkg/account"
 	"github.com/openshift-online/ocm-support-cli/pkg/capability"
 	"github.com/openshift-online/ocm-support-cli/pkg/label"
-	"github.com/openshift-online/ocm-support-cli/pkg/subscription"
 )
 
-// CmdCreateSubscriptionCapability represents the create subscription capability command
-var CmdCreateSubscriptionCapability = &cobra.Command{
-	Use:   "subscriptionCapability [subscriptionID] [capability]",
-	Short: "Assigns a Capability to a Subscription",
-	Long:  "Assigns a Capability to a Subscription",
-	RunE:  runCreateSubscriptionCapability,
+// CmdCreateAccountCapability represents the create account capability command
+var CmdCreateAccountCapability = &cobra.Command{
+	Use:   "accountCapability [accountID] [capability]",
+	Short: "Assigns a Capability to an Account",
+	Long:  "Assigns a Capability to an Account",
+	RunE:  runCreateAccountCapability,
 	Args:  cobra.ExactArgs(2),
 	PreRunE: func(cmd *cobra.Command, args []string) error {
-		subscriptionID := args[0]
+		accountID := args[0]
 		connection, err := ocm.NewConnection().Build()
 		if err != nil {
 			return fmt.Errorf("failed to create OCM connection: %v", err)
 		}
-		// validates the subscription
-		err = subscription.ValidateSubscription(subscriptionID, connection)
+		// validates the account
+		err = account.ValidateAccount(accountID, connection)
 		if err != nil {
 			return fmt.Errorf("%v", err)
 		}
 		//validates the capability
 		capabilityKey := args[1]
-		err = capability.ValidateCapability(capabilityKey, "cluster")
+		err = capability.ValidateCapability(capabilityKey, "account")
 		if err != nil {
 			return fmt.Errorf("%v", err)
 		}
@@ -41,19 +41,19 @@ var CmdCreateSubscriptionCapability = &cobra.Command{
 	},
 }
 
-func runCreateSubscriptionCapability(cmd *cobra.Command, argv []string) error {
-	subscriptionID := argv[0]
+func runCreateAccountCapability(cmd *cobra.Command, argv []string) error {
+	accountID := argv[0]
 	// TODO : avoid creating multiple connection pools
 	connection, err := ocm.NewConnection().Build()
 	if err != nil {
 		return fmt.Errorf("failed to create OCM connection: %v", err)
 	}
 	key := argv[1]
-	capabilityKey, err := capability.GetCapability(key, "cluster")
+	capabilityKey, err := capability.GetCapability(key, "account")
 	if err != nil {
 		return fmt.Errorf("failed to get capability: %v", err)
 	}
-	createdCapability, err := subscription.AddLabel(subscriptionID, capabilityKey, "true", true, connection)
+	createdCapability, err := account.AddLabel(accountID, capabilityKey, "true", true, connection)
 	if err != nil {
 		return fmt.Errorf("failed to create capability: %v", err)
 	}
