+++ cmd/ocm-support/create/accountlabel/cmd.go
@@ -1,4 +1,4 @@
-package account_label
+package accountlabel
 
 import (
 	"fmt"
