+++ cmd/ocm-support/get/organizations/cmd.go
@@ -21,17 +21,17 @@ var args struct {
 	fetchCapabilities  bool
 }
 
-// CmdFindOrganizations represents the organization find command
-var CmdFindOrganizations = &cobra.Command{
+// CmdGetOrganizations represents the organization get command
+var CmdGetOrganizations = &cobra.Command{
 	Use:   "organizations [id|external_id|ebs_account_id]",
-	Short: "Finds an organization or a list of organizations that matches the search criteria",
-	Long:  "Finds an organization or a list of organizations that matches the search criteria",
+	Short: "Gets an organization or a list of organizations that matches the search criteria",
+	Long:  "Gets an organization or a list of organizations that matches the search criteria",
 	RunE:  run,
 	Args:  cobra.ExactArgs(1),
 }
 
 func init() {
-	flags := CmdFindOrganizations.Flags()
+	flags := CmdGetOrganizations.Flags()
 	flags.BoolVar(
 		&args.all,
 		"all",
